requirements.txt
asgiref==3.6.0
channels==3.0.4
django==4.2.10
django-environ==0.11.2
djangorestframework==3.14.0
mypy==1.8.0
psycopg2-binary==2.9.9
drf-spectacular==0.26.5
django-localflavor
python-dotenv==1.0.1
channels-redis==3.4.0
djangochannelsrestframework
